migrate to new parser, add interpreter for diagnostics
